moved to appropiate header files
