+++ 2dDino.depend
@@ -916,7 +916,7 @@
 1526093416 c:\program files\codeblocks\mingw\x86_64-w64-mingw32\include\math.h
 	<crtdefs.h>
 
-1649327730 source:c:\users\samin\documents\projects\graphics\cg-project\main.cpp
+1649481274 source:c:\users\samin\documents\projects\graphics\cg-project\main.cpp
 	<time.h>
 	<cmath>
 	<math.h>
