+++ variables.h
@@ -1,0 +1,10 @@
+#ifndef VARIABLES_H_INCLUDED
+#define VARIABLES_H_INCLUDED
+
+// Player character distance from the right side of the screen
+const int x = 80;
+
+// Player character distance from the bottom of the screen
+static double w = 250;
+
+#endif // VARIABLES_H_INCLUDED
